docshell/test/mochitest/test_bug529119-2.html

<!DOCTYPE HTML>
<html>
<head>
<title>Test bug 529119</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("untriaged");

var workingURL = "http://mochi.test:8888/tests/docshell/test/mochitest/bug529119-window.html";
var faultyURL = "http://some-nonexistent-domain-27489274c892748217cn2384.com/";

var w = null;
var phase = 0;
var isWindowLoaded = false;

function delay(msec) {
  return new Promise(resolve => setTimeout(resolve, msec));
}

async function pollForPage(expectErrorPage, win) {
  // Start with polling after a delay, we might mistakenly take the current page
  // as an expected one.
  await delay(1000);

  for (let i = 0; i < 200; i++) {
    await delay(100);

    let haveErrorPage = false;
    try {
      win.document.title;
    } catch (ex) {
      haveErrorPage = true;
    }

    if (haveErrorPage == expectErrorPage) {
      return true;
    }
  }
  return false;
}

async function windowLoaded() {
  // The code under here should only be run once
  // The test popup window workingURL was already opened
  if (isWindowLoaded)
    return;
  isWindowLoaded = true;

  /* 2. We have successfully loaded a page, now go to a faulty URL */
  // XXX The test fails when we change the location synchronously
  window.setTimeout(function() {
    w.location.href = faultyURL;
  }, 0);

  ok(await pollForPage(true, w), "Waiting for error page succeeded");
  /* 3. now, while we are on the error page, navigate back */
  try {
    // We need the SpecialPowers bit, because this is a cross-origin window
    // and we normally can't touch .history on those.
    await SpecialPowers.spawn(w, [], () => this.content.history.back());
  } catch (ex) {
    ok(false, "w.history.back() threw " + ex);
  }

  ok(await pollForPage(false, w), "Waiting for original page succeeded");
  /* 4-finish, check we are back at the original page */
  is(await SpecialPowers.spawn(w, [], () => this.content.location.href),
     workingURL,
     "Is on the previous page");
  w.close();
  SimpleTest.finish();
}

function startTest() {
  /* 1. load a URL that leads to an error page */
  w = window.open(workingURL);
}

</script>
</head>
<body onload="startTest();">
</body>
</html>
